pyproject.toml
[project]
name = "django-perf-demo"
version = "1.0.0"
description = "An interactive educational tool for teaching Django ORM performance optimization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1,<6.0",
    "psycopg2-binary>=2.9.9",
    "django-debug-toolbar>=4.4.6",
    "django-silk>=5.3.2",
    "dj-database-url>=2.2.0",
    "python-decouple>=3.8",
    "factory-boy>=3.3.1",
    "Faker>=26.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]

[tool.ruff]
line-length = 88
target-version = "py312"
